requirements.txt
# 라즈베리파이 락카키 대여기 시스템 의존성

# 필수 - 시리얼 통신 (ESP32와 통신)
pyserial>=3.5

# 필수 - 웹 기반 GUI
Flask>=2.3.0
Flask-SocketIO>=5.3.0
eventlet>=0.33.0

# 필수 - 구글시트 연동 (WiFi)
gspread>=5.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# 필수 - 설정 관리
python-dotenv>=1.0.0

# 선택적 - 개발/테스트 도구
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0

# 유틸리티
psutil>=5.9.0
